Fix (one) issue with cross-module inline classes
Also, reenable KtxCrossModule tests as they seem to work
fine for me and are the only way I found to write a test for this.

Unfortunately, this does not yet allow something like `Dp` to
be an inline class. That still fails because of inlining issues
that we believe are fixed in master and should therefore
be fixed on the next compiler update. However, this issue
will still have to be fixed, so this is a step in the right
direction.

Bug: 148457543 ( https://issuetracker.google.com/issues/148457543 )
Bug: 152645654 ( https://issuetracker.google.com/issues/152645654 )
Test: KtxCrossModuleTests.testInlineClassCrossModule

Change-Id: I6c063bd74e310f41866fb9d542642c467062d422 ( https://android-review.googlesource.com/q/I6c063bd74e310f41866fb9d542642c467062d422 )

Moved from: https://github.com/androidx/androidx/commit/31aa3c95570e778982379539f9fb3792bcaec80a