Remember that the IDE is started first time, if it was restarted during config vmoptions importing: IDEA-236874
When the config is imported, the system property intellij.first.ide.session is enabled, indicating that this is the first start of this installation. Later this information is used to show the Initial Customize wizard.

However, if the vmoptions file was changed during config importing, the IDE was restarted right away, and the intellij.first.ide.session flag used to be forgotten.

To solve the issue, use the same migrate.config marker file to store system properties which should be restored after restart.

intellij-monorepo: 38e86557f2c2859cd3bc876cb9ce404f688739ba