Get rid of duplicate identical methods in class files.
The ComposableCallTransformer is introducing a new symbol for
an existing function. That leads to there being two different
symbols for the same function. In SyntheticAccessorLowering
in the Kotlin compiler, that leads to multiple accessibility
bridges. The copy is not needed.

Test: KtxCrossModuleTests:testAccessibilityBridgeGeneration
Change-Id: Id50bce6673c2ddfb77e851720031deeb919ad1eb ( https://android-review.googlesource.com/q/Id50bce6673c2ddfb77e851720031deeb919ad1eb )

Moved from: https://github.com/androidx/androidx/commit/dd97de15b6b05990163256aae7a7b1de21229081