[frameworks] don't store reference to SpringFacet in static field
This is needed to avoid failures in testDynamicExtensions after unregistering facetType extension, and storing instances from project containers in static fields is a bad practice anyway (it may lead to problems if for example loading was cancelled).

After the change in DetectedFrameworksTree there is no need to use the same instance of FrameworkType anymore.

intellij-monorepo: af264ac3821285180eea32345bb429f5e6a51baa