Include script dependencies in project resolve scopes
When a script is included in a project as a regular file to be compiled
it produces a `ModuleSourceInfo` representing its metdata, as opposed to
`ScriptModuleInfo` for the special-cased scratch files and Gradle build
script.

This change includes script dependencies as part of the regular search
scope when attempting to resolve items.