StubUpdatingIndex: on debug assertion check only that deserialized stub matches freshly built stub (attempt 2).
This commit is the second attempt of "fdc83f26: Stub indexes: move to SerializedStubTree assertion that deserialized stub of a SerializedStubTree matches original stub.", which was reverted because the added assertion fails on generating prebuilt JS indexes, for unknown reason: probably, prebuilt indexes are buggy. We had to revert that commit because it led to Installers failure (JS prebuilt indexes are generated on Installers build).

We used to have stricter assertion for non-prebuilt stubs (not only that Stubs match but also that forward index data are the same). But it is too heavy even for debug-only assertions. So we revert to the very first version of this assertion: only check that deserialized stub matches original stub (see content of this file before 29.04.2020).

intellij-monorepo: 6e6329912c6e86fa3b20429ab85bb3c5d0805759