Native: Fix invocation of the Distribution class constructor
A fix for KT-37550 requires us to stop using system properties when
the compiler is started from Gradle. In the Kotlin repository, we
stop reading the konan.home system property to determine a path to
the compiler and require this path to be explicitly specified in
the constructor of the Distribution class.

This patch changes the invocation of this constructor at the
kotlin-ultimate side.