[workspace model] WM-T-48 WM-T-47 Use same storageBuilder in `ModifiableRootModel` created from external system
For workspace model, we should de distinguish `ModifiableRootModel` created from `IdeModifiableModelsProviderImpl` or from elsewhere.
The main problem in the new model that if we will not do this we will get a lot of records with the different `entitySources` some
of them will be with `JpsImportedEntitySource` another with `FileInDirectory` but all of them should be with `JpsImportedEntitySource`.
It will lead to inconsistency during project serialization and displaying (bugs: WM-T-47, WM-T-48).

If the model created from IdeModifiableModelsProviderImpl it should use the same storage which used in the module because
modules and `ModifiableRootModels` created (`AbstractModuleDataService#createModules`) far earlier than
(`AbstractModuleDataService#setModuleOptions`) applies its changes for module's entitySource and these changes don't reflect
in created earlier ModifiableRootModel. Other entities become created with different entitySource. During commit
`AbstractIdeModifiableModelsProvider#commit` we get absolutely unusable store because the latest applied diff
will become the true state of the store.

intellij-monorepo: 603489eac7a8c9dffb75c922e1000baed25bc304