[workspace model] WM-T-41 Fix reading external system properties
Prior to this fix `LegacyBridgeModulePathMacroManager` was loaded with null as the `customModuleFilePath`
and it was lead to incorrect paths in the ExternalSystemModuleOptionsEntity in the storage. It affects on
`AutoImportProjectTracker#refreshProject ` which  execute and fails in `IdeModelsProviderImpl#isApplicableIdeModule`
comparison because of incorrect paths from `ExternalSystemModulePropertyManager#getLinkedProjectPath`.  If the
system cannot find the existing applicable module it will create the new one but with the additional prefix.
As a result, modules are duplicated

intellij-monorepo: 16a864d5f889c210e06c12840f17bad17eef6940