IDEA-CR-64251: [symbol] impl: treat self-references as declarations
This is a platform part of the fix for PY-42823. What happens there:
- `b` in `b: float = aca` is a PsiNamedElement (considered a declaration);
- `b` has a reference, which is resolved to the very same PsiNamedElement `b`;
- GTDU prefers references;
- GTDU resolves reference and navigates to the resolved PsiNamedElement effectively doing nothing.

There is a way to tell platform that some PsiNamedElement is not really a declaration by implementing
`TargetElementEvaluatorEx2#isAcceptableNamedParent`, but there was no way to tell the platform
that some PsiReference is not actually a reference but a declaration.

In old GTDU logic there was a condition that if the resolved element is the same (`==`) as the named element under caret
then Show Usages was chosen. This change introduces the same check back but on the deeper level instead of action level.

(cherry picked from commit 842503903fc1dc49fab99446494da8961dd0b51d)

intellij-monorepo: a92703fbc379f271bfa03f1edbd51ea816e90d7a