Remove the `$key` parameter from the compose ABI
This change removes the `$key` parameter as the `$key`
parameter is no longer necessary due to a change in how
source information is collected.

Also this change makes the compiler tests less noisy and
modified the restart alloaiton tracking to account for the
allocation of a source information string.

Fixes: b/158123185 ( https://issuetracker.google.com/issues/158123185 )
Test: updated tests to pass with compiler changes

Relnote: """The calling convention for composable functions has
changed. This is a binary breaking change. All libraries must be
recompiled to work with this version of the compose compiler plugin.

This change does not create a source level breaking change as the
only APIs that have changed are compiler APIs that have an explicit
opt in.
"""

Change-Id: I7afd2d7b19652ec92b8d6d1d1e92f0745968aa66 ( https://android-review.googlesource.com/q/I7afd2d7b19652ec92b8d6d1d1e92f0745968aa66 )

Moved from: https://github.com/androidx/androidx/commit/1d8872cf4f7aad93ede3277bb1627bfab188c113