Avoid registering destination dir twice
This causes a lot of tests to fail, including JS tests
that check a resulting jar not to contain a duplicate js file.
Also Android tests fail, because Dex throws an exception, when it sees duplicates

The destination dir is registered at `prepareKotlinCompileTask`,
so there is no need to register it in `registerKotlinCompileTask`
(the latter is used by `prepareKotlinCompileTask` & `registerJsLink`).