Fix testDowngradeTo106
The test was added, because switchting from 1.1 to 1.0.6 caused some problems.
There was nothing particular in 1.0.6, it simply was the previous version to 1.1,
so people naturally wanted to downgrade to 1.0.6, when they encountered issues with 1.1.

However, 1.0.6 is no longer compatible with modern versions of Gradle,
so this commit changes KGP downgrade version target to 1.3.41.