Fix compilerClasspathMutation in UpToDateIT
Previously the test printed `compileKotlin.computedCompilerClasspath`,
then set `compileKotlin.compilerClasspath` & ran a build,
then changed `compileKotlin.compilerClasspath` & ran a build again.

The problem was that `computedCompilerClasspath` was accessed during
the script evaluation (before `compilerClasspath` was set).
`computedCompilerClasspath` uses `compilerClasspath` if it is not null,
then it tries to resolve `kotlinCompilerClasspathConfiguration`.

It worked as expected when `computedCompilerClasspath` returned
a provider,which was recomputed every time.
However, during the work on supporting Gradle configuration cache
`computedCompilerClasspath` was changed to lazy, so the printing
statement broke the following changes to the `compilerClasspath`
(`compilerClasspath` was changed, but lazy `computedCompilerClasspath`
was not recomputed).

The solution is to move test specific printing in `afterEvaluate`.