Disable gradle importing tests against bootstrap `gradle-kotlin-plugin`
This commit negotiates changes which were introduced in 3 commits before.

`kotlin-ide` doesn't have neither `gradle-kotlin-plugin` sources
neither gradle which was heavily used to implement testing against
bootstrap `gradle-kotlin-plugin`. So all we got with last 3 commits
is ~100 test failures.

This commit needs to be reverted in the future when KT-40899 will be fixed

Tracking issue: KT-40899