Do not generate non-standard compareTo as primitive comparison in all backends
Previous changes related to this in the old JVM backend were in
582b1c5e66acb167b36f3ac723364ba9ecf6c836 and
0482f7e9c5627ef076f66d9fd4a2c1832446a28f, but they did not affect the
`ProperIeee754Comparisons` mode which became the default in 1.4.0. As a
result, we had a regression here.

Since the `PRIMITIVE_NUMERIC_COMPARISON_INFO` slice is used in psi2ir to
determine how to generate the comparison, this fixes the regression both
in the old JVM backend, and in all IR backends.

 #KT-41426 Fixed

(cherry picked from commit 74c6d2b9519a69818cc8b45ffbe11a6acc3f203e)