Exclude kotlin.common from big modules circular dependency
Move KotlinRoot to kotlin.common module and break circular
dependency between kotlin.common and kotlin.tests-common

This commit allows to set language level to 8 in the modules which are included into the
build process class path (KTIJ-83). Otherwise, we get "the following modules have the
same language assigned because of cyclic dependencies" error on big number of modules.