Migrate all projects with CompilationTarget.HOST to use LibraryType. Adds some more LibraryTypes. Deprecates toolingProject.
Partially inspired by Oussama's recent need to land an enforcement CL
for Lint

Test: verify that everything that used to be published is, everything
that should be CompilationTarget.HOST is, and everything that should be
API checked is. If any of these is not the case, it will trigger
warnings or errors and fail the build.

Change-Id: Ief3ffc23ea1ccf9414eb04cd641cb7a7445116d2 ( https://android-review.googlesource.com/q/Ief3ffc23ea1ccf9414eb04cd641cb7a7445116d2 )

Moved from: https://github.com/androidx/androidx/commit/74cc259900c60308eff6801cfd0d08232a6fac9e