Fix deprecated gradle api usage in artifact transform
Previously, when adding --warning-mode=fail to KaptIncrementalIT,
it would fail when testing with advance gradle version because
deprecated configurations in build files of test projects and
gradle api usage in artifact transform.

With this PR, we have two "StructureTransformAction'. The one with
latest gradle api is for use at most cases, the legacy one is for use
when gradle version is less or equal than 5.3 because the new gradle api
is introduced in 5.4.

In terms of the testing infra, this change updates deprecated
configurations to support testing artifact transform, introduces warning
mode to BuildOptions. The "--warning-mode=fail" is applied only when we
explicitly enable it in each test, gradle version is 5.6+ and agp version
3.6- is not used.

Test: KaptIncrementalIT