Disable ComposeMultiPlatformTests (2)
These tests still run despite @Ignore and not having @Test. That is
because they are JUnit3 tests, which are discovered by being methods
named `testXxx` in a subclass of `TestCase`.

Removing the test-suffix from the method names, and add an empty test
method to satisfy the requirement that a TestCase must have at least 1
test.

Bug: 171000611 ( https://issuetracker.google.com/issues/171000611 )
Test: ./gradlew :compose:compiler:compiler-hosted:integration-tests:testDebugUnitTest --rerun-tasks
Change-Id: Ida7c9271681fd74a54aeddbeba31fe5bc6a7119f ( https://android-review.git.corp.google.com/q/Ida7c9271681fd74a54aeddbeba31fe5bc6a7119f )