[FIR] fix kt-42517
After a cast on "this" of a local anonymous function,
FirDataflowAnalyzer replaces the type of the corresponding
FirThisReceiverExpressionImpl with the target type. Before the fix,
Fir2Ir conversion generates directly an IrGetValueImpl with the target
type for references to "this". During LocalDeclarationsLowering, the
anonymous function is rewritten, parameters copied including "this",
and the IrGetValueImpl replaced with a new one referring to the new
parameter whose type is the original type, not the target type (see
FunctionBodiesRewriter::visitGetValue).

Note that "this" of non-local functions or other value parameters
don't have similar problems.

There are two possible fixes:
1. Change Fir2Ir conversion to keep the cast, as in this PR;
2. Change LocalDeclarationsLowering to handle the case when
IrGetValueImpl contains an implicit cast. This option is more
intrusive than the first one.