Ban inter-library Experimental/OptIn API usage from outside library group
This doesn't check for requireSameVersion but it's a good start in
anticipation of the ability to do so.

Also we use a heuristic to determine the maven group; we take the
first two words of the package and join those with a ".", this may
not work well for maven groups with three words, but it shouldn't
result in false positives.

Bug: 162069062 ( https://issuetracker.google.com/issues/162069062 )
Test: gw lintDebug
Change-Id: I54b5d82d3b9660f45ca874cb2255c8c11eb28459 ( https://android-review.git.corp.google.com/q/I54b5d82d3b9660f45ca874cb2255c8c11eb28459 )