Make "classes" finalizedBy "ide-plugin-dependencies:jar" in cooperative compilation mode
Cooperative compilation mode is mode with src-to-src compilation with kotlin-ide repo.
When developer hits "Build project button" in his/her IDE then IntelliJ IDEA triggers
'classes' and 'testClasses' tasks in Gradle. But in order to achieve successfull
compilation in src-to-src mode we need to produce not only classes but jars as well.
(see more in kotlin-ide/kotlin/kotlinc-delegators)