Adds lint check for composable lambda parameters in composable functions
Composable functions with only one composable lambda parameter should use the name `content` and place this parameter at the end, so it can be used as a trailing lambda.

Also updates existing functions that do not follow these guidelines.

Bug: b/172469237 ( https://issuetracker.google.com/issues/172469237 )
Test: ComposableLambdaParameterDetectorTest
Test: updateApi
Test: lintDebug
Relnote: """Added lint check for composable lambda parameter naming and position, to check for consistency with Compose guidelines.
Also migrated some APIs using `children` as the name for their trailing lambda to `content`, according to the lint check and guidance."""

Change-Id: Iec48e38a2896785b521814d95c9fb624d2807315 ( https://android-review.git.corp.google.com/q/Iec48e38a2896785b521814d95c9fb624d2807315 )