[FIR] fix property accessors from fakeoverriding delegated properties.
When a property in a subclass is created from fakeoverriding a delegated property in a super class, the delegated property does not have FIR accessors, and DeleagatedMemberGenerator creates IR accessors. This causes IrProperty.discardAccessorsAccordingToBaseVisibility to discard the subclass property accessors.

This PR adds a check in IrProperty.discardAccessorsAccordingToBaseVisibility for the delegated property case, so that subclass property accessors can be kept.

^Fix KT-43342.