Fix NPE in AndroidProgramRunner.canRun() in debugger tests run against Android Studio
AndroidProgramRunner.canRun() triggers on launching the debuggee process. It expects the run configuration Project to be set, however debugger tests use MockConfiguration that unconditionally returns 'null' for it.

As we can't change Android Studio sources or easily patch class files, the work-around is not that nice, however it fixes the exception.

The fix is supposed to be put to 202 and 203 branches ONLY.

(cherry picked from commit 2e1ece2739fa947b62e2f9c7425729a20ad9c25a)